fix: fixed audio bug
